refactor(ContactsList): Move helpers to their dedicated file
